clarify invoke member function implementation
I accidentally discovered that by declaring a templated data member
pointer type, the type can be deduced to a function type, resulting in a
member function pointer type. This was I could deduce cv-qualifiers and
noexcept-ness all in one, without having to resort to more
metaprogramming or writing 8x overloads.

Removed parens and added SFINAE checks to clarify that the parameter is
a member pointer type, and the member type gets deduced to a function
type, resulting in a member function pointer.